Move network type check to inside the network code
Prevents races when the all-in-one is used with multi-tenant SDN
